Fixed a nasty timestamp bug
